locations for board mounts
The outer radius is the "platform" that goes under the
board.  The inner radius is the "cap" that clamps down on
the PCB (and which must be clear of features on the top
of the PCB).

From the drawing it looks like there are some reasonable
locations, but I need to check it visually against the
actual RPi to make sure that there is nothing protruding
on the top or bottom that would get in the way.

I feel fairly confident that in conjunction with the
standoffs below the mounting holes, I can get a pretty
secure mounting for the RP.

The 5"x2.75" form factor seems to work quite well - the
screw holes for the board mounts won't be unreasonably
close to the edge of the bottom plate.