moved board mounts to avoid hitting stuff on pcb
From visual inspection of the RPi, there are various
things on the underside of the PCB that the board mounts
would have hit in their original positions.  Moved them
to places where the top and bottom are clear.  One
issue is that the corner with the GPIO header has no
board mount - there doesn't seem to be any good place
for one.  Will double-check when the board mounts
arrive and I can try them out on the actual RPi to
see how they fit.