move standoffs and RPi for better clearance
Added a layer to the drawing to show the approximate size
of the standoffs.  Was concerned that the bottom right
standoff would block access to the ethernet port, so moved
all of the standoffs 1mm horizontally and vertically
towards the corners, and moved the RPi further towards
the top.  This should allow good clearance for the
ethernet.