starting on UI
Trying to put in a relatively GUI-agnostic framework
without making things unnecessarily complicated.  Initial
implementation is ncurses, but I want to leave open
the possibility of an actual GUI implementation in
the future.

One thing that is missing is controllers: views should
only handle events related to painting and updates of
model objects.

Things kind of work at this point: a menu is displayed,
but changing the selection doesn't work.