added MenuController; made ConsMenuView read-only
From now on, ALL of the views will be read-only, and
this will be enforced through having model object pointer
be to a const object.

This should work really well: since we have our
own event system, the views have no obligation to
handle user input events, and in fact do not even
need to be aware of the existence of the controllers.